+++ api/src/context/NoopContextManager.ts
@@ -34,9 +34,6 @@ export class NoopContextManager implements types.ContextManager {
 
   /** @experimental This API is experimental and may change in minor releases without prior notice. */
   attach(_context: types.Context): types.Token {
-    return ROOT_CONTEXT as unknown as types.Token;
+    return { dispose() {} };
   }
-
-  /** @experimental This API is experimental and may change in minor releases without prior notice. */
-  detach(_token: types.Token): void {}
 }